Update the way of checking if gerrit is ready
As of M80, native support of htmlImports is removed,
replaced with polyfill from webcomponentsjs.
One of the biggest change is switching from
sync to async which will affect quite a lot on how
we load plugins.

As Gerrit should be ready once header shows up,
here replaced the old checking on # of imports to
the readiness of #mainHeader

Change-Id: I351c4ddb4abf8fb5fb5d5f57bd5986c593357283