Migrate to Manifest V3: Migrate to a service worker
We are following this migration guide:
https://developer.chrome.com/docs/extensions/migrating/to-service-workers/

Update the "background" field in the manifest: Simple straightforward
changes in the src/manifest.json file.

Move DOM and window calls to an offscreen document: Nothing to do.

Convert localStorage to chrome.storage.local: Nothing to do.

Register listeners synchronously: `setUpListeners()` and
`removeListeners()` functions are removed. We are just always listening
to events.

Replace calls to XMLHttpRequest() with global fetch(): Nothing to do.

Persist states: The global `Map` `TabStates` are moved to
`chrome.storage.session`. Helper functions are added. async/await is
added where needed.

Convert timers to alarms: `setIntercal()` is replaced by using the
`chrome.alarms` API. We cannot check every second, but 30 seconds is
hopefully good enough. We are also adding an additional listener to
`chrome.tabs.onUpdated` to make sure that icon and popup are always
up-to-date.

Keep the service worker alive (in exceptional cases): Nothing to do.

Change-Id: I583b0ff70a414c33c66781846d866e7784460011