Migrate to Manifest V3: Migrate to a service worker
We are following this migration guide:
https://developer.chrome.com/docs/extensions/migrating/to-service-workers/

Adding a StorageUtil class for interacting with `chrome.storage`. We are
persisting the rules and the enabled tabs in storage. The service
worker, the popup and the content script can all access
`chrome.storage`, but the content script can only access
`chrome.storage.sync`, not `.session`.

Register listeners synchronously: `setUpListeners()` and
`removeListeners()` functions are removed. We are just always listening
to events.

Persist states: The global `Map` `TabStates` are moved to
`chrome.storage.session`. Helper functions are added. async/await is
added where needed.

Convert timers to alarms: Remove timer. Not needed anymore. Event
listeners are working just fine.

`checkCurrentTab()` was re-purposed to `updateIconPopup()`. Enabling and
disabling a tab happens directly by calling `storage` methods. You can
listen to `storage` changes via `chrome.storage.onChanged`. Thus we only
need a function in the service worker for updating the icon and popup
states when either the active tab changes or when the tab is enabled or
disabled.

Remove obsolete `isNew` property from `Rule` objects.

Change-Id: I583b0ff70a414c33c66781846d866e7784460011