Migrate to Manifest V3: Replace blocking web request listeners
We are following this migration guide:
https://developer.chrome.com/docs/extensions/migrating/blocking-web-requests/

In a nutshell Chrome does not allow request/response manipulation
anymore programmatically: You cannot hook up a listener and do whatever
you like adhoc. Instead you have to register declarative manipulation
rules. This suits us quite well, because we already have declarative
rules defined by the user.

So we are removing all the listeners and the adhoc manipulation, and
instead we are adding conversion logic from our own `Rule` interface
to Chrome's `chrome.declarativeNetRequest.Rule`.

Then we only have to listen to `chrome.storage` updates and call
`updateRules()`, which will do the conversion and make sure that
Chrome's rules are in sync with the user's rules.

We are also removing default rules being fetched from a URL. Instead we
believe that using the default rules that were shipped with the current
version of the extension is more robust, and simpler to maintain.

Google-Bug-Id: b/287605771
Change-Id: I92548878dd82067ed6e14cf9102d6dc51379c56d