Convert tip and snackbar to Lit elements
`INJECT_JS_CODE` is not supported anymore, so we cannot define showing
the red "helper tip" in the bottom right corner as a rule. Instead we
have to do this statically. Let's use the opportunity and do this as a
Lit component instead of using `document.createElement()` stuff.

Also convert the snackbar to a Lit component.

Unfortunately that requires us to also inject `webcomponents-lite.js`
polyfill.

Also upgrade to Lit 3 to be compatible with Gerrit's main web app.

Google-Bug-Id: b/287605771
Change-Id: I1096d5b58377d98eb54754d54f38892ee3d420c4