Continue rules processing after first disabled one
With the Manifest V3 migration in 619b1197, the logic to process rules
used a filter() with disabled rules using an empty return to be filtered
out. The refactoring removed the filter but kept the return which means
the first disabled rule causes an early return and no other rule is
processed.

Since the rules are now processed in a loop, a disabled rule should
cause a continuation of the next iteration.

Change-Id: Ieef2a3c4af2bc43b39ec71682483cb9144327f87