Fix injecting JS plugins, remove webcomponents-lite.js
Loading `webcomponents-lite.js` as a content script is problematic,
because it does not have proper access to the `window` object of the
page. It also seemed a bit ugly, because Gerrit would load it again
itself, only too late.

We are now taking a different approach. We are moving all the custom
elements that we want to use into a separate `elements.js` bundle,
and load that script "normally" from a <script> tag in the header,
exactly as we do for injected js plugins.

Change-Id: I5662989b115a57e751842f8b19afacb33433e5f1