Remove support for INJECT_JS_PLUGIN and INJECT_JS_MODULE_PLUGIN
Content Security Policy for extensions does not allow injecting
arbitrary JS.

We have considered passing a list of URLs from the extension to the
web app instead, but that is non-trivial to do, because they also
cannot access the `window` object of each other. The extension could
add an element to the DOM and set an attribute on that. The web app
would then read the attribute and load those plugins instead of the
ones listed in the server config. Apart from adding complexity the
disadvantage is that the web app would have to be changed, and thus
you could not use the concept for existing open source servers.

There is a trivial workaround though: You can just redirect an exsiting
plugin to your new plugin. Every Gerrit server typically has a plugin
that can be replaced during development, so this is not much of a
limitation.

Google-Bug-Id: b/317481553
Release-Notes: Remove INJECT_JS_PLUGIN and INJECT_JS_MODULE_PLUGIN
Change-Id: If954960ad56aba5595c6ec029ce8211fcb6ec5b0